+++ services/soso-web/src/app/layout.tsx
@@ -1,26 +1,22 @@
 // src/app/layout.tsx
+
 import "./globals.css";
 import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import { SnackbarProvider } from "@/components/ui/snackbar"; // ★追加
-
-const inter = Inter({ subsets: ["latin"] });
+import { RootLayout } from "@/layouts/AppLayout/AppLayout"; 
 
 export const metadata: Metadata = {
   title: "SOSo",
   description: "SoSo dispatch app",
 };
 
-export default function RootLayout({
+export default function AppRouterLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
   return (
     <html lang="ja">
-      <body className={inter.className}>
-        <SnackbarProvider>{children}</SnackbarProvider>
-      </body>
+      <RootLayout>{children}</RootLayout>
     </html>
   );
 }
