+++ services/soso-web/src/views/ResisterView/schema.ts
@@ -61,4 +61,4 @@ export const registerSchema = z
     }
   });
 
-export type RegisterFormValues = z.infer<typeof registerSchema>;
+export type RegisterViewValues = z.infer<typeof registerSchema>;
